slight restructuring of energy/rapl logic for zero divide

+++ src/plugins/acct_gather_energy/rapl/acct_gather_energy_rapl.c
@@ -225,7 +225,7 @@ extern int acct_gather_energy_p_update_node_energy(void)
 	acct_gather_energy_shutdown = false;
 	if (!acct_gather_energy_shutdown) {
 		uint32_t node_current_energy;
-		float node_freq;
+		uint16_t node_freq;
 
 		hardware();
 		for (i = 0; i < nb_pkg; i++)
@@ -247,11 +247,11 @@ extern int acct_gather_energy_p_update_node_energy(void)
 			local_energy->consumed_energy =
 				node_current_energy - local_energy->base_watts;
 			local_energy->current_watts =
-				(float)(node_current_energy -
-				local_energy->previous_consumed_energy);
+				node_current_energy -
+				local_energy->previous_consumed_energy;
 			node_freq = slurm_get_acct_gather_node_freq();
 			if (node_freq)	/* Prevent divide by zero */
-				local_energy->current_watts /= node_freq;
+				local_energy->current_watts /= (float)node_freq;
 		}
 		if (local_energy->consumed_energy == 0) {
 			local_energy->consumed_energy = 1;
